organize molecool into modules and subpackage
