+++ molecool/measure.py
@@ -1,0 +1,42 @@
+import numpy as np
+
+"""
+This module is for functions that perform measurements.
+"""
+
+
+def calculate_angle(rA, rB, rC, degrees=False):
+    AB = rB - rA
+    BC = rB - rC
+    theta = np.arccos(np.dot(AB, BC) / (np.linalg.norm(AB) * np.linalg.norm(BC)))
+
+    if degrees:
+        return np.degrees(theta)
+    else:
+        return theta
+
+
+def calculate_distance(rA, rB):
+    """Calculate the distance between two points.
+
+    Parameters
+    ----------
+    rA, rB : np.ndarray
+        The coordinates of each point.
+
+    Returns
+    -------
+    distance : float
+        The distance between the two points.
+
+    Examples
+    --------
+    >>> r1 = np.array([0, 0, 0])
+    >>> r2 = np.array([0, 0.1, 0])
+    >>> calculate_distance(r1, r2)
+    0.1
+    """
+    dist_vec = rA - rB
+    distance = np.linalg.norm(dist_vec)
+
+    return distance
